+++ tests/Phuria/SQLBuilder/Test/Integration/ParametersTest.php
@@ -29,7 +29,7 @@ class ParametersTest extends DatabaseTestCase
     {
         $connection = $this->createQueryConnection();
 
-        $qb = static::qbFactory()->createSelect();
+        $qb = static::phuriaSQLBuilder()->createSelect();
         $userTable = $qb->from('user');
         $qb->addSelect($userTable->column('username'));
         $qb->andWhere("{$userTable->column('id')} = :id");
@@ -46,7 +46,7 @@ class ParametersTest extends DatabaseTestCase
     {
         $connection = $this->createQueryConnection();
 
-        $qb = static::qbFactory()->createSelect();
+        $qb = static::phuriaSQLBuilder()->createSelect();
         $userTable = $qb->from('user');
         $qb->addSelect($userTable->column('username'));
         $qb->andWhere("{$userTable->column('id')} = :id");
@@ -64,7 +64,7 @@ class ParametersTest extends DatabaseTestCase
     public function itWillSelectNotExistingUser()
     {
         $connection = $this->createQueryConnection();
-        $qb = static::qbFactory()->createSelect();
+        $qb = static::phuriaSQLBuilder()->createSelect();
         $userTable = $qb->from('user');
         $qb->addSelect($userTable->column('username'));
         $qb->andWhere("{$userTable->column('id')} = :id");