+++ tests/Phuria/SQLBuilder/Test/Integration/QueryBuilder/UpdateBuilderTest.php
@@ -26,7 +26,7 @@ class UpdateBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function updateWithSetClause()
     {
-        $qb = static::qbFactory()->createUpdate();
+        $qb = static::phuriaSQLBuilder()->createUpdate();
 
         $exampleTable = $qb->update('example');
         $qb->addSet("{$exampleTable->column('name')} = NULL");
@@ -46,7 +46,7 @@ class UpdateBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function updateWithWhereClause()
     {
-        $qb = static::qbFactory()->createUpdate();
+        $qb = static::phuriaSQLBuilder()->createUpdate();
 
         $rootTable = $qb->update('example');
         $qb->addSet("{$rootTable->column('name')} = NULL");
@@ -61,7 +61,7 @@ class UpdateBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function updateWithOrderByAndLimit()
     {
-        $qb = static::qbFactory()->createUpdate();
+        $qb = static::phuriaSQLBuilder()->createUpdate();
 
         $rootTable = $qb->update('example');
         $qb->addSet("{$rootTable->column('salary')} = 100");
@@ -77,7 +77,7 @@ class UpdateBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function updateIgnore()
     {
-        $qb = static::qbFactory()->createUpdate();
+        $qb = static::phuriaSQLBuilder()->createUpdate();
 
         $qb->setIgnore(true);
         $qb->update('example');